Create MN KHAN
Create a 3D illusion picture where a boy in a black shirt sits casually on a Wingback Chair. Wearing sneakers, a black cricket cap, and sunglasses, he looks ahead. The background features “Your Name” in big and capital white fonts on the black wall. There should not be his shadow, and there are wings to make it appear as if he is an angel.